requirements.txt
appdirs==1.4.4
ariadne==0.13.0
ariadne-jwt==0.1.5
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
django-cors-headers==3.7.0
graphql-core==3.1.5
isort==5.9.3
mypy-extensions==0.4.3
pathspec==0.9.0
promise==2.3
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
starlette==0.14.2
tomli==1.2.0
typing-extensions==3.10.0.0
